#374 Title and category should be centered if no excerpt is given
