+++ app/src/main/java/it/niedermann/owncloud/notes/model/SectionViewHolder.java
@@ -9,9 +9,9 @@ import it.niedermann.owncloud.notes.databinding.ItemNotesListSectionItemBinding;
 public class SectionViewHolder extends RecyclerView.ViewHolder {
     private final ItemNotesListSectionItemBinding binding;
 
-    public SectionViewHolder(View view) {
-        super(view);
-        binding = ItemNotesListSectionItemBinding.bind(view);
+    public SectionViewHolder(ItemNotesListSectionItemBinding binding) {
+        super(binding.getRoot());
+        this.binding = binding;
     }
 
     public void bind(SectionItem item) {
