+++ app/src/main/res/layout/activity_notes_list_view.xml
@@ -124,7 +124,7 @@
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:scrollbars="vertical"
-                    tools:listitem="@layout/item_notes_list_note_item" />
+                    tools:listitem="@layout/item_notes_list_note_item_with_excerpt" />
             </FrameLayout>
         </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
