+++ app/src/main/res/layout/item_notes_list_note_item_with_excerpt.xml
