+++ app/src/main/res/layout/item_notes_list_note_item_without_excerpt.xml
@@ -1,0 +1,96 @@
+<?xml version="1.0" encoding="utf-8"?>
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/noteSwipeFrame"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:background="@color/bg_attention">
+
+    <ImageView
+        android:id="@+id/noteFavoriteLeft"
+        android:layout_width="32dp"
+        android:layout_height="32dp"
+        android:layout_gravity="center_vertical"
+        android:layout_marginStart="@dimen/button_padding"
+        android:layout_marginLeft="@dimen/button_padding"
+        android:contentDescription="@string/menu_favorite"
+        app:srcCompat="@drawable/ic_star_white_24dp" />
+
+    <ImageView
+        android:id="@+id/noteDeleteRight"
+        android:layout_width="32dp"
+        android:layout_height="32dp"
+        android:layout_gravity="end|center_vertical"
+        android:layout_marginEnd="@dimen/button_padding"
+        android:layout_marginRight="@dimen/button_padding"
+        android:contentDescription="@string/menu_delete"
+        app:srcCompat="@drawable/ic_delete_white_32dp" />
+
+    <LinearLayout
+        android:id="@+id/noteSwipeable"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:background="@drawable/list_item_background_selector"
+        android:baselineAligned="false"
+        android:gravity="center_vertical">
+
+        <FrameLayout
+            android:layout_width="wrap_content"
+            android:layout_height="match_parent">
+
+            <ImageView
+                android:id="@+id/noteFavorite"
+                android:layout_width="wrap_content"
+                android:layout_height="match_parent"
+                android:background="?attr/selectableItemBackgroundBorderless"
+                android:contentDescription="@string/menu_favorite"
+                android:padding="@dimen/spacer_2x"
+                tools:src="@drawable/ic_star_yellow_24dp" />
+
+            <androidx.appcompat.widget.AppCompatImageView
+                android:id="@+id/noteStatus"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_gravity="center_vertical|end"
+                android:layout_marginTop="12dp"
+                android:layout_marginEnd="4dp"
+                android:layout_marginRight="4dp"
+                android:baseline="14dp"
+                app:srcCompat="@drawable/ic_sync_blue_18dp" />
+        </FrameLayout>
+
+        <TextView
+            android:id="@+id/noteTitle"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            android:ellipsize="end"
+            android:maxLines="1"
+            android:paddingTop="@dimen/spacer_2x"
+            android:paddingBottom="@dimen/spacer_2x"
+            android:textColor="?android:textColorPrimary"
+            android:textSize="@dimen/primary_font_size"
+            tools:text="@tools:sample/lorem/random" />
+
+        <TextView
+            android:id="@+id/noteCategory"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="@dimen/spacer_1x"
+            android:layout_marginLeft="@dimen/spacer_1x"
+            android:layout_marginEnd="@dimen/spacer_2x"
+            android:layout_marginRight="@dimen/spacer_2x"
+            android:background="@drawable/border"
+            android:maxLines="1"
+            android:paddingLeft="@dimen/spacer_1x"
+            android:paddingRight="@dimen/spacer_1x"
+            android:paddingBottom="1dp"
+            android:singleLine="true"
+            android:textColor="?android:textColorPrimary"
+            android:textSize="@dimen/secondary_font_size"
+            tools:maxLength="15"
+            tools:text="@tools:sample/lorem/random" />
+    </LinearLayout>
+
+</FrameLayout>
